move game to modules
