+++ src/modules/Game/Game.tsx
@@ -1,9 +1,8 @@
 import React, { FunctionComponent, useState, useEffect } from "react";
 import { useTheme, Layer } from "sancho";
-import { GameMenu } from "../GameMenu";
-import { GameField } from "../GameField";
-/* import { preatyArray, randomByPercent } from "./gameFunctions"; */
-import { preatyArray, randomByPercent } from "components/Game/gameFunctions";
+import { GameMenu } from "components/GameMenu";
+import { GameField } from "components/GameField";
+import { preatyArray, randomByPercent } from "./gameFunctions";
 
 const startMock = {
   field: {
